% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_pivot_longer.R
\name{coda_pivot_longer}
\alias{coda_pivot_longer}
\title{Convert coda summary output to pivot_longer style data frame columns
One column for posterior values, one column for variable names, one column for 2.5 ci, one column for 97.5 ci}
\usage{
coda_pivot_longer(var_list, coda_sum, colnams = NULL)
}
\value{
A data frame of columns for posterior means, 2.5 \%, and 97.5 CI quantiles for each variable
}
\description{
Function to extract posterior means, and 2.5 and 97.5 CI quantiles from jags coda output.
It takes a list of variable names and the coda summary.
All variables in the list MUST have the same length posterior outputs.
Returns tidyverse pivot_longer format, best for distribution plotting and facet_grids
}
