% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_rows_to_cols.R
\name{coda_rows_to_cols}
\alias{coda_rows_to_cols}
\title{Convert coda rows to data frame columns}
\usage{
coda_rows_to_cols(var_list, coda_sum, colnams = NULL)
}
\value{
A data frame of columns for posterior means, 2.5 \%, and 97.5 CI quantiles for each variable
}
\description{
Function to extract posterior means, and 2.5 and 97.5 CI quantiles from jags coda output.
It takes a list of variable names and the coda summary.
All variables in the list MUST have the same length posterior outputs.
}
