% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateconnect.R
\name{dateconnect}
\alias{dateconnect}
\title{Organize jagsUI output as a dataframe}
\usage{
dateconnect(dfobj, datevect, datename = "date", identifier = NULL, varlist)
}
\arguments{
\item{dfobj}{Dataframe you want to connect the dates to. Likely output from dumsum}

\item{datevect}{Vector of dates, doys, or timestamps}

\item{datename}{What you want the date column to be called}

\item{identifier}{the optional identifier column name from the dfobj dataframe, how the dates will connect if the variable name is the same with a different index. If not using indices to connect dates, do not assign an identifier.}

\item{varlist}{Variable names in the dfobj you want to connect dates to}
}
\value{
A data frame with a new column of dates connected to some of the variables only.
}
\description{
Function to extract posterior means, and 2.5 and 97.5 CI quantiles from jags coda output.
It takes a list of variable names and the coda summary.
All variables in the list MUST have the same length posterior outputs.
}
