% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlowdev.R
\name{findlowdev}
\alias{findlowdev}
\title{Find the chain number with the lowest deviance based on last model run.}
\usage{
findlowdev(codaobj)
}
\arguments{
\item{codaobj}{the coda object}
}
\value{
number of lowest deviance chain
}
\description{
Function to find the chain number with the lowest deviance.
}
