% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remove_index.R
\name{get_remove_index}
\alias{get_remove_index}
\title{Get the index of variables you want to remove based on the names of variables you want to keep}
\usage{
get_remove_index(to_keep, list, type)
}
\arguments{
\item{to_keep}{string of variable names}

\item{list}{list of all parameters tracked. If type is jagsUI, this needs to be in the same order you fed the parameter names into the model}

\item{type}{rjags or jagsUI. Indicates what object type coda object you're using is. This is important because rjags alphabetizes it's parameters whereas jagsUI uses the order you fed the parameters into the model}
}
\value{
A data frame of columns for posterior means, 2.5 \%, and 97.5 CI quantiles for each variable
}
\description{
Function is meant to be used with Michael Fell's removevars function for saving initials
}
