% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remove_index.R
\name{keepvars}
\alias{keepvars}
\title{Save initials based on last model run.}
\usage{
keepvars(codaobj, to_keep, paramlist, type)
}
\arguments{
\item{codaobj}{jagsUI or rjags object}

\item{to_keep}{string of variable names}

\item{paramlist}{list of all parameters tracked. If type is jagsUI, this needs to be in the same order you fed the parameter names into the model}

\item{type}{rjags or jagsUI. Indicates what object type codaobj is}
}
\value{
saved initials for next run
}
\description{
Function to get new initials based on last run.
This is essentially a small wrapper function for Michael Fell's removevars function that reverses the input so you just need the names of the variables you want to keep, as opposed to the index of the variables you want to get rid of.
}
