% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlowdev.R
\name{savelowdev}
\alias{savelowdev}
\title{Save initials of chain with the lowest deviance based on last model run.}
\usage{
savelowdev(codaobj, to_keep, paramlist, type, just_chain_number = F)
}
\arguments{
\item{codaobj}{the coda object}

\item{to_keep}{string of variable names}

\item{paramlist}{list of all parameters tracked. If type is jagsUI, this needs to be in the same order you fed the parameter names into the model}

\item{type}{rjags or jagsUI. Indicates what object type codaobj is}

\item{just_chain_number}{if set to True, will just return the lowest deviance chain number, not the initials}
}
\value{
saved initials of chain with lowest deviance for next run OR number of lowest deviance chain if just_chain_number=T
}
\description{
Function to get new initials based on the chain with the lowest deviance from last run.
}
