% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlowdev.R
\name{findlowdev}
\alias{findlowdev}
\title{Save initials of chain with the lowest deviance based on last model run.}
\usage{
findlowdev(codaobj, to_keep, paramlist, type, just_chain_number = F)
}
\arguments{
\item{to_keep}{string of variable names}

\item{paramlist}{list of all parameters tracked. If type is jagsUI, this needs to be in the same order you fed the parameter names into the model}

\item{type}{rjags or jagsUI. Indicates what object type codaobj is}
}
\value{
saved initials of chain with lowest deviance for next run OR number of lowest deviance chain if just_chain_number=T
}
\description{
Function to get new initials based on the chain with the lowest deviance from last run.
To be used with get_remove_index.
}
